Updates
Minor refactor on some components
