+++ src/app/components/Tab.jsx
@@ -8,7 +8,7 @@ export default function Tab({ href, data, text, borderT }) {
     <Link
       href={href}
       className={`block p-5 ${border} w-full text-center text-gray-500 font-bold`}
-      onClick={() => setBorder(borderT)}
+      onFocus={() => setBorder(borderT)}
       onBlur={() => setBorder("border-b-gray border-b-2")}
     >
       {text} ({data})
